# Apply torch.compile to fused_moe/grouped_topk