Inspired by https://github.com/sgl-project/sglang/commit/1ebe1d6de5e0ce082e0be059c222baf0c5ee340a. This improves the generation latency for deepseek models due to the large number of experts. For `DeepSeek-Coder-V2-Lite-Instruct` we see an increase of 154 -> 163 tokens/s or 5% improvement at bs=1 on H100

script:
```
VLLM_MLA_DISABLE=1 python benchmarks/benchmark_latency.py --model deepseek-ai/DeepSeek-Coder-V2-Lite-Instruct --max-model-len 2048 --trust-remote-code --batch-size 1 --input-len 128 --output-len 512
```

main:
```
Avg prompt throughput: 51.1 tokens/s, Avg generation throughput: 154.8 tokens/s, Running: 1 reqs, Swapped: 0 reqs, Pending: 0 reqs, GPU KV cache usage: 0.1%, CPU KV cache usage: 0.0%.
```

this pr:
```
Avg prompt throughput: 51.2 tokens/s, Avg generation throughput: 163.8 tokens/s, Running: 1 reqs, Swapped: 0 reqs, Pending: 0 reqs, GPU KV cache usage: 0.3%, CPU KV cache usage: 0.0%.
```